requirements.txt
dhanhq==1.3.2
numpy==2.0.2
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
websockets==13.0.1
